CHDEXCHAE-64 only enable signals on 2.10+ ckan, so only run code/tests when on 2.10 ckan
