+++ ckanext/ytp/comments/tests/conftest.py
@@ -13,7 +13,7 @@ import ckan.tests.helpers as helpers
 from ckan.tests import factories
 
 from ckan.lib import uploader
-from ckanext.ytp.comments.model import init_tables as ytp_init
+from ckanext.ytp.comments import model as ytp_model
 
 fake = Faker()
 
@@ -107,10 +107,46 @@ def user_factory():
 @pytest.fixture
 def clean_db(reset_db):
     reset_db()
-    ytp_init()
+    ytp_model.init_tables()
 
 
 @pytest.fixture
 def mock_storage(monkeypatch, ckan_config, tmpdir):
     monkeypatch.setitem(ckan_config, u'ckan.storage_path', str(tmpdir))
     monkeypatch.setattr(uploader, u'get_storage_path', lambda: str(tmpdir))
+
+
+class Comment(factory.Factory):
+    """A factory class for creating ytp comment. It must accept user_id and
+    package_name, because I don't want to create extra entities in database
+    during tests"""
+
+    class Meta:
+        model = ytp_model.Comment
+
+    user_id = None
+    entity_type = "dataset"
+    entity_name = None
+
+    subject = "comment-subject"
+    comment = "comment-text"
+
+    @classmethod
+    def _build(cls, target_class, *args, **kwargs):
+        raise NotImplementedError(".build() isn't supported in CKAN")
+
+    @classmethod
+    def _create(cls, target_class, *args, **kwargs):
+        if args:
+            assert False, "Positional args aren't supported, use keyword args."
+
+        kwargs["url"] = "/{}/{}".format(kwargs["entity_type"], kwargs["entity_name"])
+
+        return helpers.call_action(
+            "comment_create", context={"user": kwargs["user_id"], "ignore_auth": True}, **kwargs
+        )
+
+
+@pytest.fixture
+def comment_factory():
+    return Comment
